wallet: move AccountsResult to deprecated
